requirements.txt
# Core framework
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
pyyaml>=6.0

# Semantic memory (tutorial retrieval)
faiss-cpu>=1.7.0
FlagEmbedding>=1.2.0
numpy>=1.24.0
